Lexes out one Atom with string value and exits.
First commit since I broke everything up in to modules and I'm lexing my
first Atom. Since it's pretty stable this is a nice spot to roll back to.

Lexer - State of the string we're currently parsing. Some helper functions
for looking this stuff up.

Grammar - Where the rules of JSON are defined. The Grammar is little
callback functions that listen out for specific characters. We then set
the attributes of the Atom accordingly.

murp.c - Where main is. Will eventually be removed from the project as it
is a lib.

README - Waffle for other folk to read to get the gist/jism.

privates - globals that only murp should know about

public - globals the user of the lib should know about