Plural elements within a Hash fixed.
Also lexer amended. I had the wrong bit of the Grammar firing off at the
end of the element. That has since been adjusted also done some sweet
jiggery.

Atom now gets passed by value to the Atomizer function. Oh yeah...
Key difference! Keys and Values are no longer dynamically allocated.

They are represented internally as Slices, a type with a pointer to a
string **and** a length for the buffer. =)

Planning on doing the same thing with the Lexer. The M.O. with the lib is
that the heap isn't mine. I get the stack, sure - but THE HEAP? Gluttony!

The reason for that is murp is kind of the anti-parser (it's a lexer I
know...) in that almost all parsers I've found online serialize to their
own custom data structures, sucking up memory along the way. This is heavy
duty, and then the user is expected to unpack on to their own types and
throw away the data the lib made. You have to wait for it to be done and
there's no room for anything ugly you might wanna write.

Ugly is powerful. Look at Perl! I love that bastard!!!