Added Probes
Not sure is different from this commit to the last since I put this down
again and I'm only tinkering because I need something to do in transit.

Probes are effectively the same as Atomizers. However there is a void
pointer included in the params. This is used for passing in any object,
the probe as it were. This is used to keep state between iterations - if
you're that way inclined.

Since it's a void * it can be any object at all, and then can be casted to
what it needs to be. Gotta love C.

Been thinking of working in a delimiter type which the Lexer will use.

We'll see how that one goes.