Changed LIST and HASH to OBJECT and ARRAY
I've decided to be a little more compliant with the RFC's - so I've
changed LIST to ARRAY and HASH to OBJECT. Not really a fix.

There's also a little work on the Container recognition. But it's nothing
super special. I'll be trying to get that done over this weekend.

As always work has been not a-happening.