Parses Containers (ARRAYS and OBJECTS)
If we find an Container in our JSON we fish it out. We do this a little
weirdly - if I do say so myself. We count up the delimiters, rather than
keeping state on a stack or whatever. This is so I don't have to malloc
anything - which goes against the grain of this project.

That means if we encounter a weird object that has the right amount of
delimiters but they're not in the right ordered we throw an error and
break out. I do think this is a little harsh and could be more flexible.

We'll see what I do about it. Also the Containers aren't checked to see
if they're valid JSON. We just rip the most valid feeling chunk out.
We would call a fresh Atomize on it to validate the object.