In the middle of upgrading the Onject parsing
Still in the middle of sussing out the new object parsing. I wouldn't
normally push something I feel is so broken but I want it up on GitHub so
I can pull it down wherever and get some work done... For some obscure
reason I've been working on this on a headless FreeBSD VM on my laptop.

Your guess is as good as mine...