Weird issues dried up on Linux
So removing debug junk. Objects now parse correctly. Arrays are next,
then numbers, then the keyword-y ones true, false and null.

Objects and Arrays aren't checked for much. Only the right amount of
delimiters, and that the delimiters are in the right order.

This is done by adding a fresh sub call on to the stack when we hit some
opening delimiter and returning only when we meet the matching delimiter.

It ensures the nests are a little more correct. It doesn't check to see if
these nests are valid JSON though. This is because a fresh Atomize/Probe
should be called on the Slice.

Technically these changes were in the last commit, but were not
logged because I just wanted to shove the code up on GitHub.