Fixed some issues with nested Containers
Spent a good long while trial and error-ing, trying to get the Containers
to parse right. I reckon this'll be the ticket now.

Fingers crossed anyway...

I wound up simplifying how we scan through the characters. It makes sense
now. Was that subtle kind of pushing and pulling on the data till it all
fell in to place. Hopefully it doesn't require more work.

We'll see. As always!