Version 1.0 of lib completed
The library has been completed. Arrays should now be parsed too.

There needs to be further testing and better documentation but we're off
to a start at least.

I've added the FreeBSD License to the README. I'll probably drop this in
to each of the modules.

How exciting! I'll try to bundle in a few more examples later...